LangChain_Models/requirements.txt
# langchain core
langchain==0.3.25
langchain-core==0.3.62

# langchain integrations
langchain-openai==0.3.18
openai==1.82.0




# Anthropic integration
langchain-anthropic==0.3.14

# Google Gemini (PaLM) integration
langchain-google-genai==2.1.5

# huggingface integration
langchain-huggingface==0.2.0
transformers==4.52.3
huggingface-hub==0.32.2

# environment variables Management
python-dotenv==1.1.0

# machine learning utilities
scikit-learn==1.6.1
numpy==2.2.6




